Flyttet formvalidering ut av komponenten
